+++ packages/utils/hooks/useNotification/index.tsx
@@ -11,9 +11,15 @@ import type { UseNotification } from './types'
 const createNotificationStore = () =>
   createStore<UseNotification>((set) => ({
     ...INITIAL_NOTIFICATION_STATE,
-    sendToast: (message, { type = 'success' } = {}) => set({ message, type, open: true }),
-    sendApiErrorToast: (error) =>
-      set({ message: getApiErrorMessage(error), type: 'error', open: true }),
+    sendToast: (message, { type = 'success', shouldShowProgress } = {}) =>
+      set({ message, type, open: true, shouldShowProgress }),
+    sendApiErrorToast: (error, { shouldShowProgress } = {}) =>
+      set({
+        message: getApiErrorMessage(error),
+        type: 'error',
+        open: true,
+        shouldShowProgress,
+      }),
     closeToast: () => set({ open: false }),
   }))