+++ packages/utils/hooks/useNotification/types.ts
@@ -1,12 +1,16 @@
 export type NotificationState = {
   open: boolean
+  shouldShowProgress?: boolean
   message: string
   type: 'success' | 'info' | 'warning' | 'error'
 }
 
 type NotificationFunctions = {
-  sendToast: (message: string, options?: { type?: NotificationState['type'] }) => void
-  sendApiErrorToast: (error: unknown) => void
+  sendToast: (
+    message: string,
+    options?: { type?: NotificationState['type']; shouldShowProgress?: boolean },
+  ) => void
+  sendApiErrorToast: (error: unknown, options?: { shouldShowProgress?: boolean }) => void
   closeToast: () => void
 }