+++ packages/components/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @baseapp-frontend/components
 
+## 1.0.4
+
+### Patch Changes
+
+- Implement a snackbar component with a 'progress bar' indicating the remaining time before it automatically disappears
+- Updated dependencies
+  - @baseapp-frontend/design-system@1.0.4
+  - @baseapp-frontend/utils@3.1.5
+  - @baseapp-frontend/authentication@4.1.6
+  - @baseapp-frontend/graphql@1.2.6
+
 ## 1.0.3
 
 ### Patch Changes